clean up compiler warnings
The augmented-assignment tests in particular generated a lot of
unused-variable and unused-assignment warnings, which are undesirable
not only on æsthetic developer-experience grounds but, more importantly,
because genuinely useful warnings inadvertently set off by future
development efforts could get lost in the noise. Typically one would
indicate an intentionally-unused variable by choosing an identifier
starting with an underscore, but in the case of the augmented-assignment
tests, the `allow`-directive attributes seem more appropriate: the
`a_slice` variables weren't _really_ unused; it's just that `rustc`
isn't smart enough to notice that they're views into an underlying
matrix which is being used.