conversions from Vector to (n-by-1) Matrix
Sometimes we want to think of (column) vectors as a separate type, but
sometimes we want to think of them as a matrix with one column. (For
example, as the test suggests, the scalar product u⃗· v⃗ can be profitably
reconceptualized as a 1-by-1 matrix u⃗ᵀv⃗.)

Put this in a new `linalg::convert` module to keep it out of the way.