bump rulinalg to ... master
The new uses of `assert_matrix_eq!` were failing because
`rulinalg::macros::matrix_eq::AbsoluteError<f64>` is private. This seems to
have been fixed in AtheMathmo/rulinalg@294c1199, which isn't in a published
version yet. For the moment, we can pin to the most recent (at commit time)
commit of rulinalg.